Fix unchecked offsets in binary mime.cache parsing
- CacheFile::getUint16()/getUint32()/getCharStar() did raw pointer
arithmetic on the mmap'd mime.cache buffer using offsets taken straight
from the file, without checking them against the mapped file's actual
size before every use (matchGlobList, matchSuffixTree,
matchMagicRule, addParents, resolveAlias, iconForMime all trust
these offsets verbatim).
A corrupted or malicious mime.cache can drive an out-of-bounds read of
the mapping, leading to a crash or, in principle, memory disclosure.
- Add checking against the mapped file's actual size.
- Add checking operations with offsets don't overflow.

Task-number: QTBUG-148640
Pick-to: 6.12 6.11

+++ src/corelib/mimetypes/qmimeprovider_p.h
@@ -107,10 +107,10 @@ private:
 
     quint32 matchGlobList(QMimeGlobMatchResult &result, CacheFile *cacheFile,
                           const RecordList &globs, const QString &fileName);
-    bool matchSuffixTree(QMimeGlobMatchResult &result, CacheFile *cacheFile, int numEntries,
-                         int firstOffset, const QString &fileName, qsizetype charPos,
+    bool matchSuffixTree(QMimeGlobMatchResult &result, CacheFile *cacheFile, quint32 numEntries,
+                         quint64 firstOffset, const QString &fileName, qsizetype charPos,
                          bool caseSensitiveCheck);
-    bool matchMagicRule(CacheFile *cacheFile, int numMatchlets, int firstOffset,
+    bool matchMagicRule(CacheFile *cacheFile, quint32 numMatchlets, quint64 firstOffset,
                         const QByteArray &data);
     QLatin1StringView iconForMime(CacheFile *cacheFile, const RecordList &icons,
                                   QStringView inputMime);
